+++ scripts/handlers/a5e.js
@@ -2,8 +2,8 @@
  * Level Up (A5E)
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 /**
  *
