+++ scripts/handlers/black-flag.js
@@ -2,8 +2,8 @@
  * Black Flag Roleplaying
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 /**
  *
