+++ scripts/handlers/cc.js
@@ -1,17 +1,17 @@
 /**
- * Forbidden Lands
+ * Castles and Crusades
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-export class ForbiddenLandsSystem {
-    static stackedItemTypes = ['consumable', 'loot', 'tool'];
+export class CCSystem {
+    static stackedItemTypes = ['Equipment:consumable'];
 
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemTypes(ForbiddenLandsSystem.stackedItemTypes);
+        Registry.registerStackedItemTypes(CCSystem.stackedItemTypes);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
