+++ scripts/handlers/demonlord.js
@@ -2,8 +2,8 @@
  * Shadow of the Demon Lord
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 /**
  *
