+++ scripts/handlers/dnd5e.js
@@ -1,17 +1,17 @@
 /**
- * Hackmaster, 5th Edition
+ * Dungeons and Dragons, 5th Edition
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-export class Hackmaster5eSystem {
+export class DnD5eSystem {
     static stackedItemTypes = ['consumable', 'loot', 'tool'];
 
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemTypes(Hackmaster5eSystem.stackedItemTypes);
+        Registry.registerStackedItemTypes(DnD5eSystem.stackedItemTypes);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
