+++ scripts/handlers/dungeonworld.js
@@ -1,17 +1,17 @@
 /**
- * Castles and Crusades
+ * Dungeonworld
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-export class CCSystem {
-    static stackedItemTypes = ['Equipment:consumable'];
+export class DungeonworldSystem {
+    static stackedItemTypes = ['consumable', 'loot', 'tool'];
 
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemTypes(CCSystem.stackedItemTypes);
+        Registry.registerStackedItemTypes(DungeonworldSystem.stackedItemTypes);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
