+++ scripts/handlers/genesys.js
@@ -1,34 +1,31 @@
 /**
- * The One Ring, 1st Edition
+ * Genesys
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-/**
- *
- */
-export class TOR1eSystem {
-    static stackedItemTypes = ['miscellaneous'];
+export class GenesysSystem {
+    static stackedItemTypes = ['consumable', 'gear'];
 
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemTypes(TOR1eSystem.stackedItemTypes);
+        Registry.registerStackedItemTypes(GenesysSystem.stackedItemTypes);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
-        Registry.registerDirectiveHandler('currency', TOR1eSystem._handleCurrency);
+        Registry.registerDirectiveHandler('currency', GenesysSystem._handleCurrency);
     }
 
     static async _handleCurrency(actor, args) {
-        Logging.debug('_handleCurrency', actor, args);
+        Logging.debug('handleCurrency', actor, args);
 
         let quantity = args.quantity;
 
-        const currentAmount = actor.system.treasure.value;
+        const currentAmount = actor.system.currency;
         Logging.debug("currentAmount", currentAmount);
         const amount = parseInt(currentAmount) + parseInt(quantity);
         Logging.debug('amount', amount);
-        await actor.update({ ['system.treasure.value']: amount });
+        await actor.update({ 'system.currency': amount });
     }
 }
