+++ scripts/handlers/hackmaster5e.js
@@ -1,17 +1,17 @@
 /**
- * Dungeons and Dragons, 5th Edition
+ * Hackmaster, 5th Edition
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-export class DnD5eSystem {
+export class Hackmaster5eSystem {
     static stackedItemTypes = ['consumable', 'loot', 'tool'];
 
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemTypes(DnD5eSystem.stackedItemTypes);
+        Registry.registerStackedItemTypes(Hackmaster5eSystem.stackedItemTypes);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
