+++ scripts/handlers/ose.js
@@ -2,8 +2,8 @@
  * Old-School Essentials
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 export class OSESystem {
     static stackedItemTypes = ['item'];
