+++ scripts/handlers/pf1.js
@@ -1,17 +1,22 @@
 /**
- * Worlds Without Number
+ * Pathfinder, 1st Edition
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-export class WWNSystem {
-    static stackedItemTypes = ['item'];
+/**
+ *
+ */
+export class PF1System {
 
+    /**
+     *
+     */
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemCallback(WWNSystem._isItemStackable);
+        Registry.registerStackedItemCallback(PF1System._isItemStackable);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
@@ -20,8 +25,10 @@ export class WWNSystem {
     static _isItemStackable(item) {
         Logging.debug("_isItemStackable", item);
 
-        return (WWNSystem.stackedItemTypes.contains(item.type) &&
-                !item.system.hasCharges);
+        return ((item.type == 'consumable') ||
+            (item.type == 'loot' &&
+                (item.system.subType == 'misc' ||
+                    item.system.subType == 'ammo')));
     }
 
 }
