+++ scripts/handlers/pf2e.js
@@ -2,8 +2,8 @@
  * Pathfinder, 2nd Edition
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 export class PF2eSystem {
     static currencySlugMap = {
