+++ scripts/handlers/shadowdark.js
@@ -2,8 +2,8 @@
  * Shadowdark RPG
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 /**
  *
