+++ scripts/handlers/t2k4e.js
@@ -2,9 +2,9 @@
  * Twilight:2000
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
-import { Utils } from "./utilities.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
+import { Utils } from "../utilities.js";
 
 /**
  *
