+++ scripts/handlers/weirdwizard.js
@@ -2,8 +2,8 @@
  * Shadow of the Weird Wizard
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 export class SotWWSystem {
     static stackedItemTypes = ['Equipment:consumable'];
