+++ scripts/handlers/wfrp4e.js
@@ -2,8 +2,8 @@
  * Warhammer Fantasy Roleplay, 4th Edition
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
 export class WFRP4eSystem {
     static stackedItemTypes = ['trapping', 'money', 'ammunition'];
