+++ scripts/handlers/wwn.js
@@ -1,22 +1,17 @@
 /**
- * Pathfinder, 1st Edition
+ * Worlds Without Number
  */
 import { AllSystems } from "./handlers-all.js";
-import { Registry } from "./registry.js";
-import { Logging } from "./logging.js";
+import { Registry } from "../registry.js";
+import { Logging } from "../logging.js";
 
-/**
- *
- */
-export class PF1System {
+export class WWNSystem {
+    static stackedItemTypes = ['item'];
 
-    /**
-     *
-     */
     static registerHandlers() {
         Logging.debug("registerHandlers");
 
-        Registry.registerStackedItemCallback(PF1System._isItemStackable);
+        Registry.registerStackedItemCallback(WWNSystem._isItemStackable);
         Registry.registerLinkEntryHandler(AllSystems.handleLinkEntry);
         Registry.registerTextEntryHandler(AllSystems.handleTextEntry);
         Registry.registerDirectiveHandler('currency', AllSystems.handleCurrency);
@@ -25,10 +20,8 @@ export class PF1System {
     static _isItemStackable(item) {
         Logging.debug("_isItemStackable", item);
 
-        return ((item.type == 'consumable') ||
-            (item.type == 'loot' &&
-                (item.system.subType == 'misc' ||
-                    item.system.subType == 'ammo')));
+        return (WWNSystem.stackedItemTypes.contains(item.type) &&
+                !item.system.hasCharges);
     }
 
 }
